refactor: separated the widgets in buy dialog into its own classes
