+++ lib/dialogs/buy/buy_dialog.dart
@@ -1,11 +1,11 @@
 import 'package:ax_dapp/athlete/athlete.dart' hide AptTypeSelectionChanged;
 import 'package:ax_dapp/dialogs/buy/bloc/buy_dialog_bloc.dart';
+import 'package:ax_dapp/dialogs/buy/widgets/widgets.dart';
 import 'package:ax_dapp/scout/models/models.dart';
 import 'package:ax_dapp/service/confirmation_dialogs/custom_confirmation_dialogs.dart';
 import 'package:ax_dapp/service/controller/controller.dart';
 import 'package:ax_dapp/service/custom_styles.dart';
 import 'package:ax_dapp/util/bloc_status.dart';
-import 'package:ax_dapp/util/helper.dart';
 import 'package:ax_dapp/util/util.dart';
 import 'package:ax_dapp/util/warning_text_button.dart';
 import 'package:ax_dapp/wallet/wallet.dart';
@@ -69,137 +69,6 @@ class _BuyDialogState extends State<BuyDialog> {
           Expanded(
             child: ShortAptButton(),
           )
-        ],
-      ),
-    );
-  }
-
-  Widget showTotalFee(String totalFee) {
-    return Flexible(
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Text(
-            'Total Fees:',
-            style: textStyle(
-              Colors.grey[600]!,
-              15,
-              isBold: false,
-              isUline: false,
-            ),
-          ),
-          Text(
-            '$totalFee AX(0.3%)',
-            style: textStyle(
-              Colors.grey[600]!,
-              15,
-              isBold: false,
-              isUline: false,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget showBalance(double balance) {
-    return Flexible(
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.end,
-        children: [
-          Text(
-            'Balance: ${toDecimal(balance, 6)}',
-            style: textStyle(
-              Colors.grey[600]!,
-              15,
-              isBold: false,
-              isUline: false,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget showMarketPriceImpact(String marketPriceImpact) {
-    return Flexible(
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Text(
-            'Market Price Impact:',
-            style: textStyle(
-              Colors.grey[600]!,
-              15,
-              isBold: false,
-              isUline: false,
-            ),
-          ),
-          Text(
-            '$marketPriceImpact %',
-            style: textStyle(
-              Colors.grey[600]!,
-              15,
-              isBold: false,
-              isUline: false,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget showMinimumReceived(String minimumReceived) {
-    return Flexible(
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Text(
-            'Minimum Received:',
-            style: textStyle(
-              Colors.grey[600]!,
-              15,
-              isBold: false,
-              isUline: false,
-            ),
-          ),
-          Text(
-            '$minimumReceived APT',
-            style: textStyle(
-              Colors.grey[600]!,
-              15,
-              isBold: false,
-              isUline: false,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget showSlippage(double slipageTolerance) {
-    return Flexible(
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Text(
-            'Slippage Tolerance:',
-            style: textStyle(
-              Colors.grey[600]!,
-              15,
-              isBold: false,
-              isUline: false,
-            ),
-          ),
-          Text(
-            '$slipageTolerance %',
-            style: textStyle(
-              Colors.grey[600]!,
-              15,
-              isBold: false,
-              isUline: false,
-            ),
-          ),
         ],
       ),
     );
@@ -457,7 +326,7 @@ class _BuyDialogState extends State<BuyDialog> {
                       Row(
                         mainAxisAlignment: MainAxisAlignment.end,
                         children: [
-                          showBalance(balance),
+                          Balance(balance: balance),
                         ],
                       ),
                     ],
@@ -480,23 +349,23 @@ class _BuyDialogState extends State<BuyDialog> {
                       Row(
                         mainAxisAlignment: MainAxisAlignment.spaceBetween,
                         children: [
-                          showTotalFee(totalFee.toStringAsFixed(6)),
+                          TotalFees(totalFee: totalFee),
                         ],
                       ),
                       Row(
                         mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                        children: [showMarketPriceImpact(priceImpact)],
+                        children: [MarketPriceImpact(priceImpact: priceImpact)],
                       ),
                       Row(
                         mainAxisAlignment: MainAxisAlignment.spaceBetween,
                         children: [
-                          showMinimumReceived(minReceived),
+                          MinimumReceived(minReceived: minReceived),
                         ],
                       ),
                       Row(
                         mainAxisAlignment: MainAxisAlignment.spaceBetween,
                         children: [
-                          showSlippage(slippageTolerance),
+                          SlippageTolerance(slippageTolerance: slippageTolerance),
                         ],
                       ),
                     ],
@@ -622,88 +491,6 @@ class ShortAptButton extends StatelessWidget {
           fontSize: 11,
         ),
       ),
-    );
-  }
-}
-
-class Price extends StatelessWidget {
-  const Price({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        Text(
-          'Price:',
-          style: textStyle(
-            Colors.white,
-            15,
-            isBold: false,
-            isUline: false,
-          ),
-        ),
-        BlocBuilder<BuyDialogBloc, BuyDialogState>(
-          buildWhen: (previous, current) =>
-              previous.aptTypeSelection != current.aptTypeSelection ||
-              previous.aptBuyInfo != current.aptBuyInfo ||
-              previous.longApt != current.longApt ||
-              previous.shortApt != current.shortApt,
-          builder: (context, state) {
-            final price = state.aptBuyInfo.axPerAptPrice.toStringAsFixed(6);
-            final _textStyle = textStyle(
-              Colors.white,
-              15,
-              isBold: false,
-              isUline: false,
-            );
-            return state.aptTypeSelection.isLong
-                ? Text(
-                    '$price AX per ${state.longApt.ticker} APT',
-                    style: _textStyle,
-                  )
-                : Text(
-                    '$price AX per ${state.shortApt.ticker} APT',
-                    style: _textStyle,
-                  );
-          },
-        )
-      ],
-    );
-  }
-}
-
-class AmountToReceive extends StatelessWidget {
-  const AmountToReceive({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    final _textStyle = textStyle(
-      Colors.white,
-      15,
-      isBold: false,
-      isUline: false,
-    );
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        Text('You Receive:', style: _textStyle),
-        BlocBuilder<BuyDialogBloc, BuyDialogState>(
-          builder: (context, state) {
-            final amountToReceive =
-                state.aptBuyInfo.receiveAmount.toStringAsFixed(6);
-            return state.aptTypeSelection.isLong
-                ? Text(
-                    '$amountToReceive ${state.longApt.ticker} APT',
-                    style: _textStyle,
-                  )
-                : Text(
-                    '$amountToReceive ${state.shortApt.ticker} APT',
-                    style: _textStyle,
-                  );
-          },
-        )
-      ],
     );
   }
 }
