+++ lib/dialogs/buy/widgets/widgets.dart
@@ -1,0 +1,7 @@
+export 'amount_to_receive.dart';
+export 'balance.dart';
+export 'market_price_impact.dart';
+export 'minimum_received.dart';
+export 'price.dart';
+export 'slippage_tolerance.dart';
+export 'total_fees.dart';
